Added support for production/staging in organisation add migration

+++ db/migrations/1725254645149-AddTrace.js
@@ -2,6 +2,8 @@ module.exports = class AddTrace1725254645149 {
   name = "AddTrace1725254645149";
 
   async up(db) {
+    const SQUID_NETWORK = process.env.SQUID_NETWORK || "eth-sepolia";
+    if (SQUID_NETWORK !== "eth-sepolia") return;
     // add organisation with name "Trace" and schema id "0x2e22df9a11e06c306ed8f64ca45ceae02efcf8a443371395a78371bc4fb6f722"
     await db.query(
       `INSERT INTO "organisation" ("id", "name", "issuer", "color") 
@@ -15,6 +17,8 @@ module.exports = class AddTrace1725254645149 {
   }
 
   async down(db) {
+    const SQUID_NETWORK = process.env.SQUID_NETWORK || "eth-sepolia";
+    if (SQUID_NETWORK !== "eth-sepolia") return;
     // remove organisation with name "Trace" and schema id "0x2e22df9a11e06c306ed8f64ca45ceae02efcf8a443371395a78371bc4fb6f722"
     await db.query(
       `DELETE FROM "organisation" WHERE "id" = '0x2e22df9a11e06c306ed8f64ca45ceae02efcf8a443371395a78371bc4fb6f722'`
