+++ db/migrations/1727410589775-AddGiveth Verification Team.js
@@ -1,7 +1,9 @@
-module.exports = class AddGivethVerificationTeam1725344741698 {
-  name = "AddGivethVerificationTeam1725344741698";
+module.exports = class AddGivethVerificationTeam1727410589775 {
+  name = "AddGivethVerificationTeam1727410589775";
 
   async up(db) {
+    const SQUID_NETWORK = process.env.SQUID_NETWORK || "eth-sepolia";
+    if (SQUID_NETWORK !== "eth-sepolia") return;
     // add organisation with name "Giveth Verification Team" and schema id "0xf63f2a7159ee674aa6fce42196a8bb0605eafcf20c19e91a7eafba8d39fa0404"
     await db.query(
       `INSERT INTO "organisation" ("id", "name", "issuer", "color") 
@@ -15,6 +17,8 @@ module.exports = class AddGivethVerificationTeam1725344741698 {
   }
 
   async down(db) {
+    const SQUID_NETWORK = process.env.SQUID_NETWORK || "eth-sepolia";
+    if (SQUID_NETWORK !== "eth-sepolia") return;
     // remove organisation with name "Giveth Verification Team" and schema id "0xf63f2a7159ee674aa6fce42196a8bb0605eafcf20c19e91a7eafba8d39fa0404"
     await db.query(
       `DELETE FROM "organisation" WHERE "id" = '0xf63f2a7159ee674aa6fce42196a8bb0605eafcf20c19e91a7eafba8d39fa0404'`
